Add ability for pluginMan to not install the .plg
Currently, to use the plugin manager as an easy method of running one-off scripts that DO NOT need to be re-run at boot time, the script needs to exit with a non-zero return code to force pluginMan to not copy the .plg to /boot/config/plugins.

The trigger for this is a new <PLUGIN> attribute of "noInstall"  When its value is non-null, the .plg will not get saved to /bot/config/plugins, but the .plg can still exit with a proper zero return code.

Admittedly, not too much call for this, but it would have come in handy for my Support / Project template update script for 6.4.1-rc2