Remove control codes, and extended ascii characters from plugin urls on install
Every once in a while, a copy / paste from a browser of CA's URL posted on the forum winds up with unicode and/or control characters in the middle of the URL, which prevents installation.  (https://lime-technology.com/forums/topic/38582-plug-in-community-applications/?page=85&tab=comments#comment-669454  and  https://lime-technology.com/forums/topic/38582-plug-in-community-applications/?page=84&tab=comments#comment-664991 )


Just replace them with a null character.

Probably geographically related.